Split up mobile_app bits into individual files
